requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
Django==4.1.6
django-crispy-forms==1.14.0
djangorestframework==3.14.0
Pillow==9.4.0
requests==2.25.1
whitenoise==6.4.0
django-storages==1.13.2
azure-core==1.26.3
azure-storage-blob==12.14.1
mysqlclient==2.0.3
mysql-connector-python